pilviotus/requirements.txt
astroid==2.4.2
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
firebase-admin==4.5.2
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-bigquery==2.8.0
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-pubsub==2.2.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
httplib2==0.19.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
libcst==0.3.16
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.2
proto-plus==1.13.0
protobuf==3.14.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3339==6.2
rsa==4.7
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.3
wrapt==1.12.1
